refactor(connlib): track UDP DNS query source in query meta data (#7018)
When performing recursive DNS queries over UDP, `connlib` needs to
remember the original source socket a particular query came from in
order to send the response back to the correct socket. Until now, this
was tracked in a separate `HashMap`, indexed by upstream server and
query ID.

When DNS queries are being retried, they may be resent using the same
query ID, causing "Unknown query" logs if the retry happens on a shorter
interval than the timeout of our recursive query.

We are already tracking a bunch of meta data along-side the actual
query, meaning we can just as easily add the original source socket to
that as well.

Once we add TCP DNS queries, we will need to track the handle of the TCP
socket in a similar manner.