fix(connlib): limit number of max buffered messages to portal (#7264)
When we have connectivity issues to the portal, such as in #7254 or
#7200, messages will pile up in this buffer. In case Firezone is left
running for long enough, the connection will fail after a certain number
of retries. Often times though, the user will exit Firezone before that
because nothing is working.

We have to mitigate unbounded buffer growth anyway and clearing the
buffer gives us an opportunity to emit a warning which will send an
alert to Sentry, notifying us that something isn't working for a
particular customer.

Resolves: #6397.