fix(connlib): apply timeout to WebSocket connection to portal (#7265)
The issue in #7254 and #7200 appears to be that eventually, we fail to
connect to the portal because we stop re-trying, i.e. the socket connect
appears to hang forever. Perhaps there is race condition somewhere in
how we resolve DNS / flush DNS servers etc. Regardless of that,
connecting to the portal should never take more than 5s so timing out
after that ensures we retry the connection.

Resolves: #7254.
Resolves: #7200.